+++ SecondExercizeApp/app/src/main/java/com/daracul/android/secondexercizeapp/ui/intro/IntroFragment.java
@@ -1,11 +1,10 @@
-package com.daracul.android.secondexercizeapp;
+package com.daracul.android.secondexercizeapp.ui.intro;
 
 import android.os.Bundle;
 import android.view.Gravity;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.view.WindowManager;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
