+++ SecondExercizeApp/app/src/main/java/com/daracul/android/secondexercizeapp/ui/intro/IntroPageAdapter.java
@@ -1,4 +1,4 @@
-package com.daracul.android.secondexercizeapp;
+package com.daracul.android.secondexercizeapp.ui.intro;
 
 import java.util.List;
