+++ SecondExercizeApp/app/src/main/res/layout-land/activity_about.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:fillViewport="true"
 
-    tools:context=".AboutActivity">
+    tools:context=".ui.about.AboutActivity">
 
     <LinearLayout
         android:layout_width="wrap_content"
