+++ SecondExercizeApp/app/src/main/res/layout/activity_about.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fillViewport="true"
-    tools:context=".AboutActivity">
+    tools:context=".ui.about.AboutActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
