+++ SecondExercizeApp/app/src/main/res/layout/activity_news_list.xml
@@ -13,7 +13,7 @@
             android:id="@+id/recycler"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            tools:context=".NewsListActivity"
+            tools:context=".ui.list.NewsListActivity"
             tools:listitem="@layout/news_item" />
 
         <com.google.android.material.floatingactionbutton.FloatingActionButton
